+++ client/src/components/pages/SignIn/SignIn.jsx
@@ -1,28 +1,18 @@
 import { useEffect } from 'react'
 import { useNavigate } from 'react-router-dom'
-import { useSelector, useDispatch } from 'react-redux'
-import { useForm } from 'react-hook-form'
-import { login } from '../../../features/auth/authSlice'
+import { useSelector } from 'react-redux'
 import ROUTES from '../../../constants/routes'
+import SignInForm from './SignInForm'
 
 function SignIn() {
-  const { isError, isSuccess, isLoading, message } = useSelector(
-    (state) => state.auth
-  )
+  const { token, isLoading, message } = useSelector((state) => state.auth)
   const navigate = useNavigate()
-  const dispatch = useDispatch()
-  const {
-    register,
-    handleSubmit,
-    formState: { errors }
-  } = useForm()
-  const onSubmit = (credentials) => dispatch(login(credentials))
 
   useEffect(() => {
-    if (isSuccess) {
+    if (token) {
       navigate(ROUTES.PROFILE)
     }
-  }, [isSuccess, navigate])
+  }, [token, navigate])
 
   return (
     <main className="main bg-dark">
@@ -30,35 +20,8 @@ function SignIn() {
         <i className="fa fa-user-circle sign-in-icon" />
         <h1>Sign In</h1>
         {isLoading && <p className="loading">Loading...</p>}
-        {isError && <p className="error">{message}</p>}
-        {isSuccess && <p className="success">{message}</p>}
-        <form onSubmit={handleSubmit(onSubmit)}>
-          <div className="input-wrapper">
-            <label htmlFor="email">Email</label>
-            <input
-              type="text"
-              id="email"
-              {...register('email', { required: true })}
-            />
-            {errors.email?.type === 'required' && 'Email is required'}
-          </div>
-          <div className="input-wrapper">
-            <label htmlFor="password">Password</label>
-            <input
-              type="password"
-              id="password"
-              {...register('password', { required: true })}
-            />
-            {errors.password?.type === 'required' && 'Password is required'}
-          </div>
-          <div className="input-remember">
-            <input type="checkbox" id="remember-me" />
-            <label htmlFor="remember-me">Remember me</label>
-          </div>
-          <button type="submit" className="sign-in-button">
-            Sign In
-          </button>
-        </form>
+        {message && <p className="error">{message}</p>}
+        <SignInForm />
       </section>
     </main>
   )
