+++ client/src/features/auth/authSlice.js
@@ -6,8 +6,6 @@ const token = JSON.parse(localStorage.getItem('token'))
 
 const initialState = {
   token: token || null,
-  isError: false,
-  isSuccess: false,
   isLoading: false,
   message: ''
 }
@@ -43,21 +41,18 @@ const authSlice = createSlice({
     },
     [login.fulfilled]: (state, action) => {
       state.token = action.payload
-      state.isSuccess = true
       state.isLoading = false
       state.message = 'Login successful'
     },
     [login.rejected]: (state, action) => {
       state.token = null
-      state.isError = true
       state.isLoading = false
       state.message = action.payload
     },
     [logout.fulfilled]: (state) => {
       state.token = null
-      state.isError = false
-      state.isSuccess = false
       state.isLoading = false
+      state.message = 'Logout successful'
     }
   }
 })
